refactor: test TaskStatus

+++ resources/views/task_statuses/index.blade.php
@@ -33,18 +33,22 @@
                             <td>{{ $status->name }}</td>
                             <td>{{ $status->created_at->format('d.m.Y') }}</td>
                             @auth()
-                                  <td>
-                                      <a class="text-red-600 hover:text-red-900"
-                                         rel="nofollow" data-method="delete"
-                                         data-confirm="{{ __('task_statuses.Confirm the deletion') }}"
-                                         href="{{ route('task_statuses.destroy', $status->id) }}">
-                                          {{ __('task_statuses.Delete') }}
-                                      </a>
-                                      <a class="text-blue-600 hover:text-blue-900"
-                                         href="{{ route("task_statuses.edit", $status->id) }}">
-                                          {{ __('task_statuses.Edit') }}
-                                      </a>
-                                  </td>
+                                <td>
+                                    @can('delete', $status)
+                                        <a class="text-red-600 hover:text-red-900"
+                                           rel="nofollow" data-method="delete"
+                                           data-confirm="{{ __('task_statuses.Confirm the deletion') }}"
+                                           href="{{ route('task_statuses.destroy', $status->id) }}">
+                                            {{ __('task_statuses.Delete') }}
+                                        </a>
+                                    @endcan
+                                    @can('update', $status)
+                                        <a class="text-blue-600 hover:text-blue-900"
+                                           href="{{ route("task_statuses.edit", $status->id) }}">
+                                            {{ __('task_statuses.Edit') }}
+                                        </a>
+                                    @endcan
+                                </td>
                             @endauth
                         </tr>
                     @endforeach
